+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Version 0.1.1
 
 * Added function shiftEstimate.
 
+* Added functions latticeSize, findLatticeTimeIndex.
+
 
 Version 0.1
 -----